added alter status chamado

+++ src/main/java/com/helpdesk/repository/ChamadoRepository.java
@@ -3,5 +3,5 @@ package com.helpdesk.repository;
 import com.helpdesk.model.Chamado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChamadoRepository extends JpaRepository<Chamado,Integer> {
+public interface ChamadoRepository extends JpaRepository<Chamado,Long> {
 }
